Make object mappers public
